+++ rocAL/rocAL/source/meta_data/coco_meta_data_reader.cpp
@@ -236,7 +236,7 @@ void COCOMetaDataReader::read_all(const std::string &path)
             while (parser.NextArrayValue())
             {
                 int id = 1, label = 0, iscrowd = 0;
-                std::array<float, 4> bbox;
+                std::array<long double, 4> bbox;
                 std::vector<float> mask;
                 if (parser.PeekType() != kObjectType)
                 {
